Add bb-iced-widgets
- A new crate to hold custom widgets/adjacent things.
- This allows simplifying the UI code.
