+++ Cargo.toml
@@ -14,7 +14,8 @@ members = [
 	"bb-flasher-dfu",
         "bb-drivelist", 
         "bb-flasher-mspm0",
-        "bb-bmap-parser"
+        "bb-bmap-parser",
+        "bb-iced-widgets"
 ]
 
 [profile.release]
