+++ bb-iced-widgets/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "bb-iced-widgets"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dependencies]
+iced = { version = "0.14", features = ["advanced", "canvas"] }
