+++ bb-imager-gui/src/ui/flash.rs
@@ -1,13 +1,12 @@
+use bb_iced_widgets::progress_circle;
 use iced::{
     Element,
     widget::{self, button},
 };
 
-use crate::{
-    BBImagerMessage, constants,
-    state::FlashingState,
-    ui::helpers::{self, ProgressCircle, VIEW_COL_PADDING, detail_entry, page_type1},
-};
+use crate::constants::{FONT_BOLD, TONGUE_ORANGE};
+use crate::ui::helpers::{self, VIEW_COL_PADDING, detail_entry, page_type1};
+use crate::{BBImagerMessage, state::FlashingState};
 
 pub(crate) fn view(state: &FlashingState) -> Element<'_, BBImagerMessage> {
     page_type1(
@@ -28,7 +27,7 @@ fn progress_view(state: &FlashingState) -> Element<'_, BBImagerMessage> {
         bb_flasher::DownloadFlashingStatus::Customizing => (0.99, "Customizing ..."),
     };
 
-    let progress = ProgressCircle::new(prog, 10.0f32, constants::TONGUE_ORANGE);
+    let progress = progress_circle(prog, 10.0f32, TONGUE_ORANGE, FONT_BOLD);
 
     let mut col = widget::column![progress, widget::text(label)];
     if let Some(x) = state.time_remaining() {
